[gh-41]: move some components to `skipasses` folder
