+++ src/app/skipasses/SkipassCard.component.js
